Rename ScreenGrabber file
