+++ src/ocr/screengrabber.h
